+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 BUG FIXES:
 
 * provider: Fix incorrect conversions between integer types (#278)
+* provider: Fail on bad credentials (#288)
 
 ## 0.5.0 (June 7th, 2022)